+++ Sources/GrAIdient/Core/Optimizer/OptimizerImpl.swift
@@ -176,8 +176,8 @@ class SGDOptimizer: OptimizerImpl
         let nbElems = weights.nbElems
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
         
         let command = MetalKernel.get.createCommand(
             "weightsSGD", deviceID: weights.deviceID
@@ -230,8 +230,8 @@ class SGDMomentumOptimizer: OptimizerImpl
         let nbElems = weights.nbElems
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
         
         let command = MetalKernel.get.createCommand(
             "weightsMomentum", deviceID: weights.deviceID
@@ -294,12 +294,12 @@ class AdamOptimizer: OptimizerImpl
     override func stepGPU(_ weights: IWeightBuffers)
     {
         let nbElems = weights.nbElems
-        let t = Double(_kernel.params.t)
+        let t = Float(_kernel.params.t)
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
-        let pT: [Float16] = [Float16(t)]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
+        let pT: [Float] = [t]
         
         let command = MetalKernel.get.createCommand(
             "weightsAdam", deviceID: weights.deviceID
@@ -366,12 +366,12 @@ class AMSGradOptimizer: OptimizerImpl
     override func stepGPU(_ weights: IWeightBuffers)
     {
         let nbElems = weights.nbElems
-        let t = Double(_kernel.params.t)
+        let t = Float(_kernel.params.t)
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
-        let pT: [Float16] = [Float16(t)]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
+        let pT: [Float] = [t]
         
         let command = MetalKernel.get.createCommand(
             "weightsAMSGrad", deviceID: weights.deviceID
@@ -449,12 +449,12 @@ class AdamRectifiedOptimizer: OptimizerImpl
     override func stepGPU(_ weights: IWeightBuffers)
     {
         let nbElems = weights.nbElems
-        let t = Double(_kernel.params.t)
+        let t = Float(_kernel.params.t)
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
-        let pT: [Float16] = [Float16(t)]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
+        let pT: [Float] = [t]
         
         let command = MetalKernel.get.createCommand(
             "weightsAdamRectified", deviceID: weights.deviceID
@@ -583,14 +583,14 @@ class AdaBoundOptimizer: BoundOptimizer
     override func stepGPU(_ weights: IWeightBuffers)
     {
         let nbElems = weights.nbElems
-        let t = Double(_kernel.params.t)
+        let t = Float(_kernel.params.t)
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
-        let pT: [Float16] = [Float16(t)]
-        let pLowerBound: [Float16] = [Float16(lowerBound!)]
-        let pUpperBound: [Float16] = [Float16(upperBound!)]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
+        let pT: [Float] = [t]
+        let pLowerBound: [Float] = [Float(lowerBound!)]
+        let pUpperBound: [Float] = [Float(upperBound!)]
         
         let command = MetalKernel.get.createCommand(
             "weightsAdaBound", deviceID: weights.deviceID
@@ -667,14 +667,14 @@ class AMSBoundOptimizer: BoundOptimizer
     override func stepGPU(_ weights: IWeightBuffers)
     {
         let nbElems = weights.nbElems
-        let t = Double(_kernel.params.t)
+        let t = Float(_kernel.params.t)
         
         let pNbElems: [UInt32] = [UInt32(nbElems)]
-        let pAlpha: [Float16] = [Float16(alpha)]
-        let pLambda: [Float16] = [lambda != nil ? Float16(lambda!) : 0.0]
-        let pT: [Float16] = [Float16(t)]
-        let pLowerBound: [Float16] = [Float16(lowerBound!)]
-        let pUpperBound: [Float16] = [Float16(upperBound!)]
+        let pAlpha: [Float] = [Float(alpha)]
+        let pLambda: [Float] = [lambda != nil ? Float(lambda!) : 0.0]
+        let pT: [Float] = [t]
+        let pLowerBound: [Float] = [Float(lowerBound!)]
+        let pUpperBound: [Float] = [Float(upperBound!)]
         
         let command = MetalKernel.get.createCommand(
             "weightsAMSBound", deviceID: weights.deviceID
